Added 3rd test and re-ordered malloc and status

+++ Assignment 3/my_queue.c
@@ -92,6 +92,8 @@ void display() {
   }
 }
 
-size_t returnNumberElements(){
+int returnNumberElements() {
+    //printf("\n\n%d", numberElements);
+    //display();
     return numberElements;
 }
